refactor(file watcher): merged fs & fileWatcher, updated all relevant code accordingly
